Data Compression Doc
In this article, I've explained the details of data compression and the types of data compression that should be used for various content types such as music, video, and text. Data compression is divided into two categories: lossless and lossy. In short, lossless should be used for text and written files because missing letters can cause ambiguity. However, lossless can also be used for music and video depending on the situation. I've explained this in detail with examples.